% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble.R
\name{plot_ensemble}
\alias{plot_ensemble}
\title{Title}
\usage{
plot_ensemble(
  truth,
  tibble_pred,
  incorrect = FALSE,
  tibble_prob = NULL,
  order = NULL
)
}
\arguments{
\item{order}{}
}
\description{
Title
}
